move meUserQuery, split query.ts
