+++ packages/sdk/src/graphql/clinical-api/fragment.ts
@@ -1,0 +1,9 @@
+import { graphql } from './gql';
+
+graphql(`
+  fragment SupervisorCard on Supervisor {
+    id
+    fullName
+    npi
+  }
+`);
