+++ packages/sdk/src/graphql/clinical-api/index.ts
@@ -1,0 +1,2 @@
+export * from './query';
+export * from './mutation';
