+++ packages/sdk/src/graphql/clinical-api/mutation.ts
@@ -1,0 +1,9 @@
+import { graphql } from './gql';
+
+export const CreateSupervisorMutation = graphql(`
+  mutation CreateSupervisorMutation($fullName: String!, $npi: String!) {
+    createSupervisor(input: { fullName: $fullName, npi: $npi }) {
+      ...SupervisorCard
+    }
+  }
+`);
