+++ packages/sdk/src/graphql/clinical-api/query.ts
@@ -19,25 +19,22 @@ export const SearchTreatmentOptionsQuery = graphql(`
   }
 `);
 
-graphql(`
-  fragment SupervisorCard on Supervisor {
-    id
-    fullName
-    npi
+export const MeUserQuery = graphql(`
+  query MeUserQuery {
+    me {
+      name {
+        title
+      }
+      address {
+        state
+      }
+    }
   }
 `);
 
 export const SupervisorsQuery = graphql(`
   query SupervisorsQuery {
     supervisors {
-      ...SupervisorCard
-    }
-  }
-`);
-
-export const CreateSupervisorMutation = graphql(`
-  mutation CreateSupervisorMutation($fullName: String!, $npi: String!) {
-    createSupervisor(input: { fullName: $fullName, npi: $npi }) {
       ...SupervisorCard
     }
   }
