+++ packages/sdk/src/lib.ts
@@ -22,7 +22,7 @@ import pkg from '../package.json';
 export type { LoginOptions } from './auth';
 export * as types from './types';
 export * as fragments from './fragments';
-export * from './graphql/clinical-api/query';
+export * from './graphql/clinical-api';
 export * from './graphql/clinical-api/gql';
 export type {
   FieldCompletionSnapshot,
